Add coordintorlayout for homepage

+++ domain/src/main/java/com/example/domain/repository/NoteRepository.kt
@@ -1,6 +1,7 @@
 package com.example.domain.repository
 
 import com.example.domain.model.Note
+import io.reactivex.Completable
 import io.reactivex.Single
 import io.reactivex.subjects.BehaviorSubject
 
@@ -11,7 +12,7 @@ interface NoteRepository {
 
     fun createNewNote(noteText : String) : Single<Note>
     fun getNotes() : Single<MutableList<Note>>
-    fun deleteNote(noteId: Int) : Single<Unit>
+    fun deleteNote(noteId: Long) : Single<Unit>
     fun getNoteById(noteId: Long) : Single<Note>
 
 }
